requirements.txt
discord.py
requests
python-nozomi
bs4